pyproject.toml
[tool.poetry]
name = "urls"
version = "0.1.0"
description = ""
authors = ["Renan Andrade <Fel9.renan02@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
